+++ backend/canisters/community/api/src/updates/create_channel.rs
@@ -1,7 +1,7 @@
 use candid::CandidType;
 use serde::{Deserialize, Serialize};
 use types::{
-    AccessGate, AccessRules, Avatar, ChannelId, FieldTooLongResult, FieldTooShortResult, GroupPermissions, GroupSubtype,
+    AccessGate, AccessRules, ChannelId, Document, FieldTooLongResult, FieldTooShortResult, GroupPermissions, GroupSubtype,
     Milliseconds,
 };
 
@@ -12,7 +12,7 @@ pub struct Args {
     pub description: String,
     pub rules: AccessRules,
     pub subtype: Option<GroupSubtype>,
-    pub avatar: Option<Avatar>,
+    pub avatar: Option<Document>,
     pub history_visible_to_new_joiners: bool,
     pub permissions: Option<GroupPermissions>,
     pub events_ttl: Option<Milliseconds>,