+++ backend/canisters/group/impl/src/model/mod.rs
@@ -1,1 +1,2 @@
 pub mod new_joiner_rewards;
+pub mod upgrade_instruction_counts;