+++ backend/canisters/group/impl/src/updates/remove_participant.rs
@@ -25,9 +25,9 @@ async fn remove_participant(args: Args) -> Response {
     remove_participant_impl(args.user_id, false).await
 }
 
-async fn remove_participant_impl(user_id: UserId, block: bool) -> Response {
+async fn remove_participant_impl(user_to_remove: UserId, block: bool) -> Response {
     // Check the caller can remove the user
-    let prepare_result = match read_state(|state| prepare(user_id, state)) {
+    let prepare_result = match read_state(|state| prepare(user_to_remove, state)) {
         Ok(ok) => ok,
         Err(response) => return response,
     };
@@ -35,8 +35,8 @@ async fn remove_participant_impl(user_id: UserId, block: bool) -> Response {
     // If the user is an owner of the group then call the local_user_index
     // to check whether they are a "platform moderator" in which case this removal
     // is not authorized
-    if prepare_result.is_user_an_owner {
-        match lookup_user(user_id.into(), prepare_result.local_user_index_canister_id).await {
+    if prepare_result.is_user_to_remove_an_owner {
+        match lookup_user(user_to_remove.into(), prepare_result.local_user_index_canister_id).await {
             Ok(user) if !user.is_platform_moderator => (),
             Ok(_) | Err(LookupUserError::UserNotFound) => return NotAuthorized,
             Err(LookupUserError::InternalError(error)) => return InternalError(error),
@@ -44,16 +44,16 @@ async fn remove_participant_impl(user_id: UserId, block: bool) -> Response {
     }
 
     // Remove the user from the group
-    mutate_state(|state| commit(user_id, block, prepare_result.removed_by, state))
+    mutate_state(|state| commit(user_to_remove, block, prepare_result.removed_by, state))
 }
 
 struct PrepareResult {
     removed_by: UserId,
     local_user_index_canister_id: CanisterId,
-    is_user_an_owner: bool,
+    is_user_to_remove_an_owner: bool,
 }
 
-fn prepare(user_id: UserId, state: &RuntimeState) -> Result<PrepareResult, Response> {
+fn prepare(user_to_remove: UserId, state: &RuntimeState) -> Result<PrepareResult, Response> {
     if state.data.is_frozen() {
         return Err(ChatFrozen);
     }
@@ -63,18 +63,18 @@ fn prepare(user_id: UserId, state: &RuntimeState) -> Result<PrepareResult, Respo
     if let Some(member) = state.data.get_member(caller) {
         if member.suspended.value {
             Err(UserSuspended)
-        } else if member.user_id == user_id {
+        } else if member.user_id == user_to_remove {
             Err(CannotRemoveSelf)
         } else {
             // Check if the caller is authorized to remove the user
-            let is_user_an_owner = match state.data.chat.members.get(&user_id) {
+            let is_user_to_remove_an_owner = match state.data.chat.members.get(&user_to_remove) {
                 None => return Err(UserNotInGroup),
                 Some(member_to_remove) => {
                     if member
                         .role
                         .can_remove_members_with_role(member_to_remove.role, &state.data.chat.permissions)
                     {
-                        member.role.is_owner()
+                        member_to_remove.role.is_owner()
                     } else {
                         return Err(NotAuthorized);
                     }
@@ -84,7 +84,7 @@ fn prepare(user_id: UserId, state: &RuntimeState) -> Result<PrepareResult, Respo
             Ok(PrepareResult {
                 removed_by: member.user_id,
                 local_user_index_canister_id: state.data.local_user_index_canister_id,
-                is_user_an_owner,
+                is_user_to_remove_an_owner,
             })
         }
     } else {
@@ -92,16 +92,20 @@ fn prepare(user_id: UserId, state: &RuntimeState) -> Result<PrepareResult, Respo
     }
 }
 
-fn commit(user_id: UserId, block: bool, removed_by: UserId, state: &mut RuntimeState) -> Response {
-    match state.data.chat.remove_member(removed_by, user_id, block, state.env.now()) {
+fn commit(user_to_remove: UserId, block: bool, removed_by: UserId, state: &mut RuntimeState) -> Response {
+    match state
+        .data
+        .chat
+        .remove_member(removed_by, user_to_remove, block, state.env.now())
+    {
         group_chat_core::RemoveMemberResult::Success => {
-            state.data.remove_principal(user_id);
+            state.data.remove_principal(user_to_remove);
 
             handle_activity_notification(state);
 
             // Fire-and-forget call to notify the user canister
             remove_membership_from_user_canister(
-                user_id,
+                user_to_remove,
                 removed_by,
                 block,
                 state.data.chat.name.clone(),
@@ -120,7 +124,7 @@ fn commit(user_id: UserId, block: bool, removed_by: UserId, state: &mut RuntimeS
 }
 
 fn remove_membership_from_user_canister(
-    user_id: UserId,
+    user_to_remove: UserId,
     removed_by: UserId,
     blocked: bool,
     group_name: String,
@@ -134,7 +138,7 @@ fn remove_membership_from_user_canister(
         public,
     };
     fire_and_forget_handler.send(
-        user_id.into(),
+        user_to_remove.into(),
         "c2c_remove_from_group_msgpack".to_string(),
         serialize_then_unwrap(args),
     );