+++ backend/canisters/group_index/api/src/main.rs
@@ -1,9 +1,11 @@
 use candid_gen::generate_candid_method;
 
 fn main() {
+    generate_candid_method!(group_index, active_groups, query);
     generate_candid_method!(group_index, filter_groups, query);
     generate_candid_method!(group_index, recommended_groups, query);
     generate_candid_method!(group_index, search, query);
+    generate_candid_method!(group_index, search_v2, query);
 
     generate_candid_method!(group_index, delete_frozen_group, update);
     generate_candid_method!(group_index, freeze_group, update);