+++ backend/canisters/group_index/api/src/queries/c2c_active_groups.rs
@@ -1,0 +1,2 @@
+pub type Args = crate::active_groups::Args;
+pub type Response = crate::active_groups::Response;