+++ backend/canisters/group_index/api/src/queries/mod.rs
@@ -1,4 +1,7 @@
+pub mod active_groups;
+pub mod c2c_active_groups;
 pub mod c2c_filter_groups;
 pub mod filter_groups;
 pub mod recommended_groups;
 pub mod search;
+pub mod search_v2;