+++ backend/canisters/group_index/api/src/updates/c2c_create_group.rs
@@ -1,6 +1,6 @@
 use candid::CandidType;
 use serde::{Deserialize, Serialize};
-use types::{AccessGate, AccessRules, Avatar, ChatId, GroupPermissions, GroupSubtype, Milliseconds};
+use types::{AccessGate, AccessRules, ChatId, Document, GroupPermissions, GroupSubtype, Milliseconds};
 
 #[derive(CandidType, Serialize, Deserialize, Debug)]
 pub struct Args {
@@ -9,7 +9,7 @@ pub struct Args {
     pub description: String,
     pub rules: AccessRules,
     pub subtype: Option<GroupSubtype>,
-    pub avatar: Option<Avatar>,
+    pub avatar: Option<Document>,
     pub history_visible_to_new_joiners: bool,
     pub permissions: Option<GroupPermissions>,
     pub events_ttl: Option<Milliseconds>,