+++ backend/canisters/group_index/impl/src/queries/mod.rs
@@ -1,4 +1,6 @@
+pub mod active_groups;
 pub mod filter_groups;
 pub mod http_request;
 pub mod recommended_groups;
 pub mod search;
+pub mod search_v2;