+++ backend/libraries/types/src/group_match.rs
@@ -1,4 +1,4 @@
-use crate::ChatId;
+use crate::{ChatId, CommunityId};
 use candid::CandidType;
 use serde::{Deserialize, Serialize};
 
@@ -8,4 +8,13 @@ pub struct GroupMatch {
     pub name: String,
     pub description: String,
     pub avatar_id: Option<u128>,
+}
+
+#[derive(CandidType, Serialize, Deserialize, Debug)]
+pub struct CommunityMatch {
+    pub id: CommunityId,
+    pub name: String,
+    pub description: String,
+    pub avatar_id: Option<u128>,
+    pub banner_id: Option<u128>,
 }